move from playground to v1 package + fix tests
