+++ src/main/scala/com/cognite/sdk/scala/v1/wellDataLayer.scala
@@ -1,4 +1,4 @@
-package com.cognite.sdk.scala.playground
+package com.cognite.sdk.scala.v1
 
 final case class Source(
     name: String,
